Create CellCount_defunct
File to hold the now defunct version of the Cell Count function. This version allows the user to specify a cell size in degrees but because the Earth is a spheroid the actual area of the cells varies with latitude. This is particularly pronounced at high latitudes. If only concerned with equitorial regions this is still a good approximation of equal area cells.